+++ app/routes/components/Tag/index.tsx
@@ -1,15 +1,12 @@
-import type { Route } from "./+types/index";
-
 import { ComponentTemplate } from "src/components/ComponentTemplate";
 import { Section } from "src/components/Section";
-
-import { sourceCode } from "./source_code.codegen";
-import { demoSourceCode } from "./Demo_source_code.codegen";
+import type { Route } from "./+types/index";
 import { changelog } from "./changelog.codegen";
-
 import { Demo } from "./Demo";
+import { demoSourceCode } from "./Demo_source_code.codegen";
+import { sourceCode } from "./source_code.codegen";
 
-export function meta({}: Route.MetaArgs) {
+export function meta(_: Route.MetaArgs) {
   return [
     { title: "Tag | Copy UI" },
     { name: "description", content: "The Tag component from Copy UI." },
@@ -19,12 +16,11 @@ export function meta({}: Route.MetaArgs) {
 export default function TagPage() {
   return (
     <ComponentTemplate component="Tag">
-      <Section
-        title="Tag"
-        demoAndCode={[<Demo />, demoSourceCode]}
-        sourceCode={sourceCode}
-        changelog={changelog}
-      />
+      <Section.Root title="Tag">
+        <Section.Demo node={<Demo />} code={demoSourceCode} />
+        <Section.SourceCode files={sourceCode} />
+        <Section.Changelog changelog={changelog} />
+      </Section.Root>
     </ComponentTemplate>
   );
 }