+++ src/components/Tag/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-09-26: Initial version.
+- 2025-10-08: Add alpha transparency, code formatting improvements.